feat(cli): describe orchestration flags
Fixes #17391
