+++ src/cli/command-spec.ts
@@ -9,6 +9,7 @@ export type CommandSpec = {
   summary: string
   usage: string
   allowedFlags: string[]
+  flagHelp?: Record<string, string>
   positionalArgs?: string[]
   examples?: string[]
   notes?: string[]
