+++ src/cli/handlers/orchestration-worker-cli.test.ts
@@ -741,8 +741,8 @@ describe('orchestration worker-start CLI contract', () => {
       parseArgs(['orchestration', 'worker-list', '--include-remote']).flags.get('include-remote')
     ).toBe(true)
     expect(listSpec?.allowedFlags).toContain('include-remote')
-    expect(formatCommandHelp(listSpec!)).toContain(
-      '--include-remote      Include connected-server worker observations'
+    expect(formatCommandHelp(listSpec!)).toMatch(
+      /--include-remote\s+Include connected-server worker observations/
     )
 
     callMock.mockResolvedValue({
